Restructure repo to accommodate larger number of source files arrange by purpose
