+++ src/utils/defineRibbonVals.js
