+++ src/utils/filterNewAdditionalEntries.js
